feat: add mapping to add human user from intent
